login functionality with routes + cleaned up prints in server
