+++ client/src/App.jsx
@@ -1,70 +1,14 @@
-import { useState, useEffect } from 'react'
-import loginService from './services/login'
-import LoginForm from './components/LoginForm'
+import {BrowserRouter as Router, Routes, Route} from 'react-router-dom'
+import Login from './views/login/Login'
 
 const App = () => {
-  const [errorMessage, setErrorMessage] = useState(null)
-  const [username, setUsername] = useState('') 
-  const [password, setPassword] = useState('')
-  const [user, setUser] = useState(null)
-
-  useEffect(() => {
-    const loggedUserJSON = window.localStorage.getItem('loggedInUser')
-    if (loggedUserJSON) {
-      const user = JSON.parse(loggedUserJSON)
-      setUser(user)
-    }
-  }, [])
-
-  const handleLogin = async (event) => {
-    event.preventDefault()
-    
-    try {
-      const user = await loginService.login({
-        username, password,
-      })
-
-      window.localStorage.setItem(
-        'loggedInUser', JSON.stringify(user)
-      ) 
-
-      setUser(user)
-      setUsername('')
-      setPassword('')
-    } catch (exception) {
-      setErrorMessage('Wrong credentials')
-      setTimeout(() => {
-        setErrorMessage(null)
-      }, 5000)
-    }
-  }
-
-  const loginForm = () => {
-    return(
-      <div>
-        <LoginForm
-            username={username}
-            password={password}
-            handleUsernameChange={({ target }) => setUsername(target.value)}
-            handlePasswordChange={({ target }) => setPassword(target.value)}
-            handleSubmit={handleLogin}
-          />
-      </div>
-    )
-  }
-
   return (
     <>
-      <div>
-        <h1>Welcome to LocalPro Connect</h1>
-
-        {user === null ?
-          loginForm() :
-          <div>
-            <p>{user.name} logged-in</p>
-          </div>
-        }
-      </div>
+      <Router>
+        <Routes>
+          <Route exact path='/' element={<Login/>} />
+        </Routes>
+      </Router>
     </>
   )
 }
